Move the rejection wording somewhere both lowerings can reach
The production table is about CFITSIO's grammar, not about either lowering, and
`eval::lower` needs it as it takes over validation from the arena builder. No
behaviour change: the corpus reports the same messages for every expression.
